+++ src/Vend.php
@@ -23,13 +23,12 @@ use angellco\vend\widgets\FastFeed;
 use angellco\vend\widgets\FullFeed;
 use Craft;
 use craft\base\EagerLoadingFieldInterface;
+use craft\base\Element;
 use craft\base\Field;
 use craft\base\Plugin;
 use craft\base\PreviewableFieldInterface;
 use craft\commerce\elements\Order;
-use craft\commerce\elements\Product;
 use craft\commerce\elements\Variant;
-use craft\elements\Entry;
 use craft\events\RegisterComponentTypesEvent;
 use craft\events\RegisterUrlRulesEvent;
 use craft\events\RegisterUserPermissionsEvent;
@@ -43,6 +42,7 @@ use craft\helpers\StringHelper;
 use craft\helpers\UrlHelper;
 use craft\log\FileTarget;
 use craft\services\Dashboard;
+use craft\services\Elements;
 use craft\services\UserPermissions;
 use craft\web\UrlManager;
 use venveo\oauthclient\events\TokenEvent;
@@ -270,37 +270,30 @@ class Vend extends Plugin
             );
         }
 
-        // Bind to product save event so we can update stock of composite
+        // Bind to element after save event so we can update stock of composite
         // products if need be
         Event::on(
-            Product::class,
-            Product::EVENT_AFTER_SAVE,
+            Elements::class,
+            Elements::EVENT_AFTER_SAVE_ELEMENT,
             function(Event $e) {
-                /** @var Product $product */
-                $product = $e->sender;
+                /** @var Element $element */
+                $element = $e->element;
 
-                // This is inefficient - duped code from webhooks and we could
-                // store a flag on the product itself so we can just check that
-                // rather than making another query
+                if (is_a($element, Variant::class)) {
+                    $compositeParentProductIds = Json::decode($element->vendCompositeParentProductIds);
+                    if (is_array($compositeParentProductIds)) {
+                        foreach ($compositeParentProductIds as $compositeParentProductId) {
+                            // Get the stock
+                            $stock = $this->products->calculateInventoryForComposite($compositeParentProductId);
 
-                // FECK: need to store flag on each product that is part of a bundle...
-                // json of bundle products that _use_ this product
+                            // Update the parent Entry
+                            $this->products->updateInventoryForEntry($compositeParentProductId, $stock);
 
-                // Then, on save check if its part of a bundle, then run the
-                // inventory for that whole bundle.
-
-                $entry = Entry::findOne([
-                    'vendProductId' => $product->vendProductId,
-                    'section' => 'vendProducts',
-                ]);
-
-                $productJson = Json::decode($entry->vendProductJson);
-                $compositeJson = Json::decode($entry->vendProductComposites);
-                if ($productJson['is_composite'] === true && !empty($compositeJson)) {
-                    Vend::$plugin->products->updateInventoryForCompositeProductEntry($entry);
-                    // TODO: update the variant too in another service method
+                            // Update the parent Variant
+                            $this->products->updateInventoryForVariant($compositeParentProductId, $stock);
+                        }
+                    }
                 }
-
             }
         );
 
@@ -444,9 +437,7 @@ class Vend extends Plugin
 
                     // Trigger inventory
                     foreach ($feeds->getFeeds() as $feed) {
-
                         if (StringHelper::contains($feed->feedUrl, 'vend/products/inventory')) {
-
                             $processedElementIds = [];
 
                             $queue->delay(0)->push(new FeedImport([
@@ -464,6 +455,29 @@ class Vend extends Plugin
 
                 // Main inventory feed
                 } elseif (StringHelper::contains($currentFeed->feedUrl, 'vend/products/inventory')) {
+
+                    // Trigger composites
+                    $runQueue = false;
+                    foreach ($feeds->getFeeds() as $feed) {
+                        if (StringHelper::contains($feed->feedUrl, 'vend/products/composites')) {
+                            $processedElementIds = [];
+
+                            $queue->delay(0)->push(new FeedImport([
+                                'feed' => $feed,
+                                'limit' => null,
+                                'offset' => null,
+                                'processedElementIds' => $processedElementIds,
+                            ]));
+
+                            $runQueue = true;
+                        }
+                    }
+
+                    if ($runQueue) {
+                        $queue->run();
+                    }
+
+                } elseif (StringHelper::contains($currentFeed->feedUrl, 'vend/products/composites')) {
 
                     // Trigger all of the full product import feeds
                     $runQueue = false;